requirements.txt
transformers
datasets
scipy
sentencepiece
gdown
tqdm
ftfy
regex
git+https://github.com/openai/CLIP.git
open-clip-torch
opencv-python
opencv-python-headless
openai
einops
pandas
omegaconf
iopath
fairscale
scikit-learn
# salesforce-lavis